==> bedassle2/7-evaluate_bedassle.Rmd <==
---
title: "BEDASSLE results"
output: 
  html_notebook:
    df_print: paged
---

We used SLURM to run bedassle for all species in parallel. Now we will load the results and work with the inferences.

In particular, we will evaluate the runs to check if they converged.
Then we will make IBD plots for all species, including model predictions.
Finally we will obtain the posterior distribution of $\alpha_{diff} = \alpha_p - \alpha_g$, which shows, across species, the relative importance of geography and host plant.

Finally, we will run a hierarchical model in STAN to test if the distribution of $\alpha_{diff}$can be explained by variation in species interactions. We will also test the power of this test by relabeling species to check if we could detect this with our sample sizes.

```{r}
dyn.load('~/labdir/working/inaturalist_flowers/v8/lib/libv8.so')
rm(list=ls())
library(rstan)
library(rstantools)
library(tidyverse)
library(broom)
library(ggthemes)
library(cowplot)
library(colorspace)
library(ggnewscale)
```

# Load HMC results

We will first run one example for **Anchylorhynchus trapezicollis** OTU 1 and then apply the same procedures as a function to all species.


Let's load the rstan model for **Anchylorhynchus trapezicollis** OTU 1:

```{r}
species = 'Anchylorhynchus_1'
load(str_c('bedassle_results/',species,'_model.fit.Robj'))
```

Each model can be diagnosed interactively with `shinystan::launch_shinystan(model.fit)`. We will do that below for all species, and for now will proceed to summarizing the posterior distributions by their mean and 95% credibility interval.

```{r}
estimates = summary(model.fit, pars=c('alphaE','alphaD','alpha2'))$summary %>%
  as.data.frame() %>%
  rownames_to_column('parameter') %>%
  dplyr::select(parameter,mean,`2.5%`,`97.5%`) %>%
  mutate_at(vars(-parameter),sprintf,fmt='%0.4f') %>%
  mutate(posterior=str_c(mean,' (',`2.5%`,'--',`97.5%`,')')) %>%
  select(parameter, posterior)

estimates
```

Now let's change parameter names to reflect to which distance they refer to. We will start by making a named vector to translate names

```{r}
translation = c('alpha2'='$\alpha_2$',
                'alphaD'='$\alpha_{geo}$',
                'alphaE[1]' = '$\alpha_{plant}$') 
```

Now it is easy to use this vector.

```{r}
estimates = estimates %>%
  mutate(parameter = translation[parameter])
```

Now that we know how this works, let's do it for all species. We will also format numbers as scientific notation.

```{r}
get_posteriors = function(species){
  
  load(str_c('bedassle_results/',species,'_model.fit.Robj'))
  
  
  
  estimates = summary(model.fit, 
                      pars=intersect(c('alphaE[1]',
                                       'alphaE[2]',
                                       'alpha0',
                                       'alphaD',
                                       'alpha2'),
                                     names(model.fit))) %>%
    .[['summary']] %>%
    as.data.frame() %>%
    rownames_to_column('parameter') %>%
    dplyr::select(parameter,mean,`2.5%`,`97.5%`) %>%
    mutate(mean = ifelse(parameter %in% c('alpha0', 'alpha2'),
                         sprintf(mean, fmt='%0.2f'),
                         sprintf(mean, fmt='%0.2e')),
           `2.5%` = ifelse(parameter %in% c('alpha0', 'alpha2'),
                         sprintf(`2.5%`, fmt='%0.2f'),
                         sprintf(`2.5%`, fmt='%0.2e')),
           `97.5%` = ifelse(parameter %in% c('alpha0', 'alpha2'),
                         sprintf(`97.5%`, fmt='%0.2f'),
                         sprintf(`97.5%`, fmt='%0.2e')),
           ) %>%
    mutate(posterior=str_c(mean,' (',`2.5%`,'--',`97.5%`,')')) %>%
    dplyr::select(parameter, posterior)
  
  translation = c('alpha2'='$\\alpha_2$',
                  'alpha0'='$\\alpha_0$',
                  'gamma'='$\\gamma$',
                  'alphaD'='$\\alpha_{geo}$',
                  'alphaE[1]' = '$\\alpha_{plant}$')
  
  return(estimates %>%
           mutate(parameter = translation[parameter]))
}
```

Now let's use `purrr` to run this for all species. Let's first create a vector with all species

```{r}
spp = list.files(path='bedassle_results/',pattern='.*_data.*') %>%
  str_remove('_data.*')

names(spp) = spp

spp
```

Before getting the estimates, let's make sure that the HMC ran well, using shinystan to diagnose the runs. The interactive shinystan page will open in a separate window. 

```{r}
#for (sp in spp){
#  rm(model.fit)
#  load(str_c('bedassle_results/',sp,'_model.fit.Robj'))
#  shinystan::launch_shinystan(model.fit)
#}

```

We can also summarize divergences and other problems for each run:

```{r}
for (sp in spp){
  cat('=========================\n')
  cat(sp)
  cat('\n')
  rm(model.fit)
  load(str_c('bedassle_results/',sp,'_model.fit.Robj'))
  rstan::check_hmc_diagnostics(model.fit)
  rm(model.fit)
}
```

Now let's get a dataframe with estimates for all species:

```{r}
all_spp_par = purrr::map_df(spp,get_posteriors,.id = 'species')
all_spp_par
```

# Make supplementary table

Finally, we just need to reshape the data frame to put it in the paper. Let's save it as a csv file for further editing.
```{r}
bedassle_table = all_spp_par %>%
  spread(key=parameter,value=posterior)

bedassle_table

write_csv(bedassle_table,'bedassle_results/final_table.csv')
```

# Make IBD figure

Now let's make a figure with pairwise pi against geographical, plant and climate distances for all species. 

To do that, we will load the data used in bedassle (pairwise pi and distances). Let's start with allele frequencies to obtain sample genetic covariances.
```{r}
add_dimnames = function(x, x_names){
  dimnames(x) = x_names
  return(x)
}

load(file = 'weevils_freq_data.Rdata')

ppi1 = purrr::map_df(BEDASSLE_data,~bedassle::freqs2pairwisePi(.x$freqs/2) %>%
                           add_dimnames(dimnames(.x$freqs)) %>%
                           as.dist %>%
                           broom::tidy(diagonal=F, upper=F),
                           .id = 'species') %>%
  rename(sample1 = item1,
         sample2 = item2,
         pi = distance)

ppi1
```

Now let's load data on the other distances and combine in this table. 

```{r}
pops = readr::read_csv('weevil_data.csv',guess_max = 1000) %>%
  select(genomics, population = pop_id_thesis) %>%
  distinct 


ppi = ppi1 %>% 
  left_join(pops, by = c('sample1'='genomics')) %>%
  rename(pop1 = population) %>%
  left_join(pops, by = c('sample2'='genomics')) %>%
  rename(pop2 = population)


load(file = 'plant_distances.Rdata')
load(file = 'worldclim/distances.Rdata')

ppi =  ppi %>%
  left_join(geo_dist %>%
              as.dist() %>%
              broom::tidy(diagonal=T, upper=F),
            by = c('pop1'='item1',
                   'pop2'='item2')) %>%
  rename(geo_dist=distance)

ppi =  ppi %>%
  left_join(plant_dist %>%
              as.dist() %>%
              broom::tidy(diagonal=T, upper=F),
            by = c('pop1'='item1',
                   'pop2'='item2')) %>%
  rename(plant_dist=distance)


ppi

```

Now let's reshape the table to use ggplot2 for plotting.

```{r}
ppi = ppi %>%
  gather(key='variable',value='distance',ends_with('dist')) %>%
  filter(!is.na(distance))

ppi
```

Before plotting, let's add information about insect-plant interactions
```{r}
ecology = data.frame(species = spp,
                     interaction= c(rep('Brood pollinator',3),
                                    rep('Dead tissue breeder', 6),
                                    rep('Live tissue breeder', 2),
                                    'Dead tissue breeder',
                                    'Live tissue breeder'
                                    ))

ppi = ppi %>% 
  left_join(ecology)

ppi
```

Before plotting, let's order factors so that plot panels are in the order we want them. Unfortunately, I found no way to make facet labels in ggplot that have only some part in italic and have a line break at the same time. The solution here will be to make everything italic and then manually correct in Adobe Illustrator.
```{r}
ppi = ppi %>%
  mutate(variable = factor(variable,
                           ordered=T,
                           levels=c('geo_dist','plant_dist'))) %>%
  mutate(interaction = factor(interaction,
                              ordered=T,
                              levels=c('Brood pollinator', 
                                       'Live tissue breeder',
                                       'Dead tissue breeder'))) %>%
  mutate(species = factor(species,
                          ordered = T,
                          levels = c(
                     'Anchylorhynchus_1'  ,
                     'Anchylorhynchus_2'  ,
                     'Anchylorhynchus_3'  ,
                     'M_bondari_1'        ,
                     'M_bondari_2'        ,
                     'M_ypsilon'          ,
                     'R_rectinasus_1'     ,
                     'Andranthobius_1'    ,
                     'Andranthobius_2'    ,
                     'C_decolor_1'        ,
                     'C_decolor_2'        ,
                     'C_impar'            ,
                     'D_polyphaga'        ,
                     'P_cocoseae'         
                     ), 
                     labels = c(
                     "Anchylorhynchus\ntrapezicollis OTU 1" ,
                     "Anchylorhynchus\ntrapezicollis OTU 2",
                     "Anchylorhynchus\ntrapezicollis OTU 3" ,
                     "Microstrates\nbondari OTU 1"         ,
                     "Microstrates\nbondari OTU 2"         ,
                     "Microstrates\nypsilon"               ,
                     "Remertus\nrectinasus"            ,
                     "Andranthobius\nbondari OTU 1"       ,
                     "Andranthobius\nbondari OTU 2"       ,
                     "Celetes\ndecolor OTU 1"         ,
                     "Celetes\ndecolor OTU 2"         ,
                     "Celetes\nimpar"                 ,
                     "Dialomia\npolyphaga"             ,
                     "Phytotribus\ncocoseae"              
                     )
                     )
  )
```


Now that we have the data in the format we need, let's plot.
```{r}
var_labels = c('geo_dist' = 'Geography (km)',
               'plant_dist' = 'Plant genetic distance')


p = ggplot(ppi) +
  geom_point(aes(x=distance, y=pi, color=interaction),size=0.5) +
  #scale_color_manual(values = c('Brood pollinator' = '#47DD8E',
  #                     'Live tissue breeder' = '#DD4756',
  #                     'Dead tissue breeder' = '#DD9C47'), name='Interaction') +
  scale_color_brewer(type='qual', name = 'Interaction') +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 4)) +
  facet_grid(species~variable,
             margins=F,
             scales = 'free',
             labeller = labeller(variable=var_labels, label_value),
             switch='x') +
  xlab('Distance') +
  ylab(expression(Pairwise~pi)) +
  theme_few() +
  theme(strip.text.y = element_text(angle = 0, 
                                    hjust = 0, 
                                    face = 'italic',
                                    margin = margin(l = 0)), 
        strip.placement = "outside",
        legend.position = 'bottom',
        text = element_text(size=6),
        panel.spacing.y = unit(0.1, "lines"),
        panel.spacing.x = unit(0.1, "lines"),
        legend.margin=margin(0,0,0,0),
        legend.box.margin=margin(-13,-12,-11,-12)) 

print(p)

ggsave(filename = './figures/gen_cov.pdf',
       plot = p, 
       device = 'pdf',
       width = 11.4, 
       height=12,
       units = 'cm')
  
```

To complete the plot, we will now add the average marginal effect of each distance as inferred by BEDASSLE. 

The first step is to tabulate estimated parameters for each species.

```{r}
get_average_estimates = function(species){
  
  load(str_c('bedassle_results/',species,'_model.fit.Robj'))
  load(str_c('bedassle_results/',species,'_data.block.Robj'))
  
  
  
  estimates = summary(model.fit, 
                      pars=intersect(c('gamma',
                                       'alpha0',
                                       'alphaE[1]',
                                       'alphaE[2]',
                                       'alphaD',
                                       'alpha2'),
                                     names(model.fit))) %>%
    .[['summary']] %>%
    as.data.frame() %>%
    rownames_to_column('parameter') %>%
    dplyr::select(parameter,mean)

  translation = c('gamma' = 'gamma',
                  'alpha2'='alpha_2',
                  'alphaD'='alpha_g',
                  'alpha0'='alpha_0',
                  'alphaE[1]' = 'alpha_p')
  norms = list('norm_g' = data.block$sd.geoDist,
               'norm_p' = data.block$sd.envDist[[1]])
  
  estimates = estimates %>%
           mutate(parameter = translation[parameter]) %>%
    bind_rows(as.data.frame(norms) %>%
                gather(key = 'parameter',
                       value = 'mean'))
  
  return(estimates)
  
  
}

bedassle_averages = purrr::map_df(spp,get_average_estimates,.id = 'species') %>%
  mutate(parameter = factor(parameter,
                            levels=c("alpha_0",
                                     "alpha_2",
                                     "alpha_g",
                                     "alpha_p",
                                     "norm_g",
                                     "norm_p" )
                            )) %>%
  complete(species,parameter, fill=list(mean = 0))

bedassle_averages 
```

Let's reshape it to have parameters as columns.

```{r}
bedassle_averages = bedassle_averages %>%
  spread(parameter,mean)

bedassle_averages
```

Now let's calculate the expected covariance decay for each species and distance. We will use the same formula that bedassle uses internally.

$$
Cov=\alpha_0*exp(-(\alpha_d*D)^{\alpha_2})
$$

After obtaining genetic covariance, we will calculate pairwise $\pi$. For that, we will use the reverse function that bedassle uses internally:
```{r}
bedassle::pwp2allelicCov
```

The reverse, therefore, is:
$$
  \pi=\frac{1}{2}-2*Cov
$$

```{r}
geo_df = split(bedassle_averages, bedassle_averages$species) %>%
  purrr::map_df(~data.frame(
    distance = seq(0,1,length.out = 100) * max(geo_dist), 
    covariance = .x$alpha_0 * exp(
    -(seq(0,1,length.out = 100) / .x$norm_g * .x$alpha_g)^.x$alpha_2)
    ),
    .id = 'species') %>%
  mutate(variable = 'geo_dist',
         pi = 1/2 - 2*covariance)

plant_df = split(bedassle_averages, bedassle_averages$species) %>%
  purrr::map_df(~data.frame(
    distance = seq(0,1,length.out = 100) * max(plant_dist), 
    covariance= .x$alpha_0 * exp(
    -(seq(0,1,length.out = 100) / .x$norm_p * .x$alpha_p)^.x$alpha_2)
    ),
    .id = 'species') %>%
  mutate(variable = 'plant_dist',
         pi = 1/2 - 2*covariance)

# clim_df = split(bedassle_averages, bedassle_averages$species) %>%
#   purrr::map_df(~data.frame(
#     distance = seq(0,1,length.out = 100) * max(clim_dist), 
#     covariance= .x$alpha_0 * exp(
#     -(seq(0,1,length.out = 100) / .x$norm_c * .x$alpha_c)^.x$alpha_2)
#     ),
#     .id = 'species') %>%
#   mutate(variable = 'clim_dist',
#          pi = 1/2 - 2*covariance)
    
    

curves_df = bind_rows(geo_df, plant_df) %>%
  left_join(ecology) %>%
  mutate(variable = factor(variable,
                           ordered=T,
                           levels=c('geo_dist','plant_dist'))) %>%
  mutate(interaction = factor(interaction,
                              ordered=T,
                              levels=c('Brood pollinator', 
                                       'Live tissue breeder',
                                       'Dead tissue breeder'))) %>%
  mutate(species = factor(species,
                          ordered = T,
                          levels = c(
                     'Anchylorhynchus_1'  ,
                     'Anchylorhynchus_2'  ,
                     'Anchylorhynchus_3'  ,
                     'M_bondari_1'        ,
                     'M_bondari_2'        ,
                     'M_ypsilon'          ,
                     'R_rectinasus_1'     ,
                     'Andranthobius_1'    ,
                     'Andranthobius_2'    ,
                     'C_decolor_1'        ,
                     'C_decolor_2'        ,
                     'C_impar'            ,
                     'D_polyphaga'        ,
                     'P_cocoseae'         
                     ), 
                     labels = c(
                     "Anchylorhynchus\ntrapezicollis OTU 1" ,
                     "Anchylorhynchus\ntrapezicollis OTU 2",
                     "Anchylorhynchus\ntrapezicollis OTU 3" ,
                     "Microstrates\nbondari OTU 1"         ,
                     "Microstrates\nbondari OTU 2"         ,
                     "Microstrates\nypsilon"               ,
                     "Remertus\nrectinasus"            ,
                     "Andranthobius\nbondari OTU 1"       ,
                     "Andranthobius\nbondari OTU 2"       ,
                     "Celetes\ndecolor OTU 1"         ,
                     "Celetes\ndecolor OTU 2"         ,
                     "Celetes\nimpar"                 ,
                     "Dialomia\npolyphaga"             ,
                     "Phytotribus\ncocoseae"              
                     )
                     )
  )
  
ggplot(curves_df) +
  geom_line(aes(x=distance,y=pi)) +
  facet_grid(species~variable, 
             scales = 'free',
             labeller = labeller(variable=var_labels, 
                                 species=label_value))

```


Finally, let's plot the curves together with points:

```{r}
p = ggplot(ppi) +
  geom_point(aes(x=distance, y=pi, color=interaction),size=0.5) +
  geom_line(data = curves_df,
            aes(x=distance,y=pi),size=0.2,linetype='dashed') +
  scale_color_brewer(type='qual', name = 'Interaction') +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 3),
                     labels = function(x) sprintf("%.2f", x)) +
  facet_grid(species~variable,
             margins=F,
             scales = 'free',
             labeller = labeller(variable=var_labels, 
                                 species=label_value),
             switch='x') +
  xlab('Distance') +
  ylab(expression(Pairwise~pi)) +
  theme_few() +
  theme(strip.text.y = element_text(angle = 0, 
                                    hjust = 0, 
                                    face = 'italic',
                                    margin = margin(l = 0)), 
        strip.placement = "outside",
        legend.position = 'bottom',
        text = element_text(size=6),
        panel.spacing.y = unit(0.1, "lines"),
        panel.spacing.x = unit(0.1, "lines"),
        legend.margin=margin(0,0,0,0),
        legend.box.margin=margin(-13,-12,-11,-12)) 

print(p)

ggsave(filename = './figures/gen_cov.pdf',
       plot = p, 
       device = 'pdf',
       width = 11.4, 
       height=12,
       units = 'cm')
```

Let's also try making a more horizontal figure to fit the page better. For that, we will split the figure in two and plot side by side with `cowplot`. We will also change genetic distance to % and geographical distance to thousand km to make tick labels smaller.

```{r}

clrs = RColorBrewer::brewer.pal(3,'Accent')
clrs_light = colorspace::lighten(clrs,0.5)


names(clrs) = c('Brood pollinator', 'Live tissue breeder', 'Dead tissue breeder')

var_labels = c('clim_dist' = 'Precipitation\n(mm)',
                             'geo_dist' = 'Geography\n(x1,000 km)',
                             'plant_dist' = 'Plant genetic\ndistance (%)')

ppi1 = ppi %>% 
  filter(interaction %in% c('Brood pollinator', 'Live tissue breeder'))  %>%
  mutate(species = fct_drop(species)) %>%
  mutate(distance = ifelse(variable == 'geo_dist', distance/1000, distance),
         distance = ifelse(variable == 'plant_dist', distance*100, distance))

curves1 = curves_df %>%
  filter(species %in% unique(ppi1$species)) %>%
  mutate(distance = ifelse(variable == 'geo_dist', distance/1000, distance),
         distance = ifelse(variable == 'plant_dist', distance*100, distance))

p1 = ggplot(ppi1) +
  geom_point(aes(x=distance, y=pi, color=interaction),size=0.5) +
  geom_line(data = curves1,
            aes(x=distance,y=pi),
            size=0.2,
            linetype='dashed') +
  scale_color_manual(values = clrs, 
                     name = 'Interaction', 
                     drop = FALSE,
                     guide = 'none') +
                     #guide = guide_legend(keywidth = unit(0.1,"line"))) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 3),
                     labels = function(x) sprintf("%.2f", x)) +
  scale_x_continuous(labels = function(x) sprintf("%.2f", x)) +
  facet_grid(species~variable,
             margins=F,
             scales = 'free',
             labeller = labeller(variable=var_labels, 
                                 species=label_value),
             switch='x') +
  xlab(NULL) +
  ylab(expression(Pairwise~pi)) +
  theme_few() +
  theme(strip.text.y = element_text(angle = 0, 
                                    hjust = 0, 
                                    face = 'italic',
                                    margin = margin(l = 0, r = 0)), 
        strip.text.x = element_text(margin = margin(t=0)),
        strip.placement = "outside",
        legend.position = 'bottom',
        legend.justification = 0,
        text = element_text(size=7),
        axis.text.x = element_text(hjust = 1,angle = 45),
        axis.text.y = element_text(size=6),
        plot.margin = margin(b=1.67,t=0.2,l=0.3,r=0.3, unit = 'cm'),
        panel.spacing.y = unit(0.1, "lines"),
        panel.spacing.x = unit(0.1, "lines")) 

print(p1)


ppi2 = ppi %>% 
  filter(interaction == 'Dead tissue breeder') %>%
  mutate(species = fct_drop(species)) %>%
  mutate(distance = ifelse(variable == 'geo_dist', distance/1000, distance),
         distance = ifelse(variable == 'plant_dist', distance*100, distance))

curves2 = curves_df %>%
  filter(species %in% levels(ppi2$species)) %>%
  mutate(distance = ifelse(variable == 'geo_dist', distance/1000, distance),
         distance = ifelse(variable == 'plant_dist', distance*100, distance))

p2 = ggplot(ppi2) +
  geom_point(aes(x=distance, y=pi, color=interaction),size=0.5) +
  geom_line(data = curves2,
            aes(x=distance,y=pi),
            size=0.2,
            linetype='dashed') +
  scale_color_manual(values = clrs, name = 'Interaction', guide = 'none') +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 3),
                     labels = function(x) sprintf("%.2f", x)) +
  scale_x_continuous(labels = function(x) sprintf("%.2f", x)) +
  facet_grid(species~variable,
             margins=F,
             scales = 'free',
             labeller = labeller(variable=var_labels, 
                                 species=label_value),
             switch='x') +
  xlab(NULL) +
  ylab(NULL) +
  theme_few() +
  theme(strip.text.y = element_text(angle = 0, 
                                    hjust = 0, 
                                    face = 'italic',
                                    margin = margin(l = 0,r=0)), 
        strip.text.x = element_text(margin = margin(t=0)),
        strip.placement = "outside",
        legend.position = 'bottom',
        text = element_text(size=7),
        axis.text.x = element_text(hjust = 1,angle = 45),
        plot.margin = margin(b=0.5,t=0.2,l=0.3,r=0.3, unit = 'cm'),
        #strip.text.x = element_blank(),
        axis.text.y = element_text(size=6),
        panel.spacing.y = unit(0.1, "lines"),
        panel.spacing.x = unit(0.1, "lines")) 

print(p2)

p = cowplot::plot_grid(p1,
                       p2,
                       ncol = 2,
                       rel_widths = c(0.52,0.48))
print(p)

#ggsave(filename = './figures/gen_cov_split.pdf',
#       plot = p, 
#       device = 'pdf',
#       width = 18, 
#       height=10,
#       units = 'cm')

```

Now as last step, let's create a legend for colors.

```{r}
key_df = data.frame(Pollinator = c('Non-pollinator', 'Pollinator'),
           Live.tissue = c('Dead tisue', 'Live tissue')) %>%
  expand.grid() %>%
  mutate(color= c("#FDC086",NA,"#BEAED4","#7FC97F"))
  
  
clr_key = ggplot(key_df) +
  theme_few() +
  geom_point(aes(x=Pollinator, y = Live.tissue, color = color), size = 3) +
  geom_tile(aes(x=Pollinator, y = Live.tissue),color='gray50', size = 0.5, fill = NA) +
  ggtitle("Interaction") +
  ylab('Larval breeding') +
  xlab('Adult pollination') +
  coord_equal(expand=F,ratio = 0.4) +
  scale_color_identity(guide = 'none') +
  theme(text = element_text(size=8),
        axis.title = element_blank(),
        plot.title = element_text(size=6,margin = margin(0,0,0,0)),
        plot.margin = margin(b=0.5,t=0,l=0.3,r=0, unit = 'cm'))

clr_key

ggsave('figures/color_key.pdf',
       clr_key + theme(plot.margin = margin(0.5,0.5,0.5,0.5,'mm')),
       device = 'pdf',
       width = 4*1.25, 
       height = 3*1.25,
       units = 'cm',
       useDingbats=F)

clr_grob = cowplot::as_grob(clr_key)
```


Now let's generate one of the figures for the main text. Italics will have to manually corrected in Adobe Illustrator later.
```{r}
p1 = p + cowplot::draw_grob(clr_grob,
                            width = 0.21,
                            height = 0.21,
                            x = 0.005,
                            #x = 0.318,
                            y = -0.04,
                            hjust = 0,
                            vjust = 0
                            )

print(p1)

ggsave(filename = './figures/gen_cov_split.pdf',
       plot = p1, 
       device = 'pdf',
       width = 16, 
       height= 10,
       units = 'cm',
       useDingbats = F)

```



# Testing overall effect


## Fit meta-analysis model
In addition to the plots on individual species, let's summarize, for each species, the posterior distribution of $\alpha_{dif}$, which is the difference in contribution between plant genetic distance and geography. 

Let's first define a function that pulls posterior draws of $\alpha_p$ and $\alpha_g$

```{r}
post_diff = function(species){
  
  load(str_c('bedassle_results/',species,'_model.fit.Robj'))
  
  estimates = rstan::extract(model.fit, 
                      pars=intersect(c('alphaE[1]',
                                       'alphaD'),
                                     names(model.fit))) %>%
    as_tibble %>%
    rename(aE=1, aD=2)
 
  
  return(estimates)
}
```

Now let's apply this function, calculate the difference and add information about species.
```{r}
pardifs = purrr::map_df(spp,post_diff,.id = 'species') %>%
  mutate(dif = aE-aD) %>%
  left_join(ecology) %>%
  mutate(interaction = factor(interaction,
                              ordered=T,
                              levels=c('Brood pollinator', 
                                       'Live tissue breeder',
                                       'Dead tissue breeder'))) %>%
  mutate(species = factor(species,
                          ordered = F,
                          levels = c(
                            'Anchylorhynchus_1'  ,
                            'Anchylorhynchus_2'  ,
                            'Anchylorhynchus_3'  ,
                            'M_bondari_1'        ,
                            'M_bondari_2'        ,
                            'M_ypsilon'          ,
                            'R_rectinasus_1'     ,
                            'Andranthobius_1'    ,
                            'Andranthobius_2'    ,
                            'C_decolor_1'        ,
                            'C_decolor_2'        ,
                            'C_impar'            ,
                            'D_polyphaga'        ,
                            'P_cocoseae')
  ))

```

Now let's summarize confidence intervals to plot posterior distributions.
```{r}
dif_sums = pardifs %>%
  group_by(species) %>%
  summarise(ave = mean(dif),
            l50 = quantile(dif,0.25),
            u50 = quantile(dif,0.75),
            l95 = quantile(dif,0.025),
            u95 = quantile(dif,0.975),
            interaction = unique(interaction)) %>%
  mutate(species = factor(species,
                          levels = c(
                            'Anchylorhynchus_1'  ,
                            'Anchylorhynchus_2'  ,
                            'Anchylorhynchus_3'  ,
                            'M_bondari_1'        ,
                            'M_bondari_2'        ,
                            'M_ypsilon'          ,
                            'R_rectinasus_1'     ,
                            'Andranthobius_1'    ,
                            'Andranthobius_2'    ,
                            'C_decolor_1'        ,
                            'C_decolor_2'        ,
                            'C_impar'            ,
                            'D_polyphaga'        ,
                            'P_cocoseae'),
                          labels = c(
                            "italic(Anchylorhynchus~trapezicollis)~OTU~1" ,
                            "italic(Anchylorhynchus~trapezicollis)~OTU~2",
                            "italic(Anchylorhynchus~trapezicollis)~OTU~3" ,
                            "italic(Microstrates~bondari)~OTU~1"         ,
                            "italic(Microstrates~bondari)~OTU~2"         ,
                            "italic(Microstrates~ypsilon)"               ,
                            "italic(Remertus~rectinasus)"            ,
                            "italic(Andranthobius~bondari)~OTU~1"       ,
                            "italic(Andranthobius~bondari)~OTU~2"       ,
                            "italic(Celetes~decolor)~OTU~1"         ,
                            "italic(Celetes~decolor)~OTU~2"         ,
                            "italic(Celetes~impar)"                 ,
                            "italic(Dialomia~polyphaga)"             ,
                            "italic(Phytotribus~cocoseae)"              
                          )),
         species = fct_reorder(species, ave),
         pos = ifelse(ave>0, -0.35, 0.35),
         just = ifelse(ave>0, 0, 1)) 

dif_sums
```

Let's plot the distribution of $\alpha_{diff}$, ordering species by the mean estimate. We will use here the same color key as in the last plot.

```{r}
p1 = ggplot(dif_sums) +
  geom_abline(intercept = 0,slope = 0, color = 'gray90') +
  geom_linerange(aes(x = species, ymin = l95, ymax = u95, color = interaction),
                 show.legend = F) +
  scale_color_manual(values = clrs_light, name = 'Interaction') +
  ggnewscale::new_scale_color()+
  geom_linerange(aes(x = species, ymin = l50, ymax = u50, color = interaction),
                 size=1.3,
                 show.legend = F) +
  geom_point(aes(x = species, y = ave, color = interaction),
             fill = 'white',
             stroke = 1.5,
             shape = 21,
             size = 1.5) +
  scale_color_manual(values = clrs, name = 'Interaction', guide = 'none') +
  scale_fill_manual(values = clrs_light, name = 'Interaction', guide = 'none') +
  #geom_errorbar(aes(x = species, ymin = ave, ymax=ave, color = interaction)) +
  #geom_text(aes(x=species,y=pos,label=species,hjust=just), parse=T, size = 3) +
  #guides(color = guide_legend(override.aes = list(size = 1.5))) +
  ylab(expression(alpha[diff])) +
  coord_flip()+
  xlab("OTU") +
  scale_x_discrete(labels = parse(text = levels(dif_sums$species))) +
  ylim(c(-0.35,0.35)) +
  theme_few() +
  theme(legend.position = c(0.99,0.01),
        legend.justification = c(1,0),
        text=element_text(size=9),
        legend.text = element_text(size=6),
        legend.key.size = unit(8, 'pt'),
        legend.title = element_text(size=7)) 

print(p1)
  
```


It does not look like there is a strong association between mode of interaction and whether plants drive diversification. Let's use a meta-analysis model to test whether being a pollinator decreases this difference and being an antagonist increases it. For each species, let's record the average $\alpha_{diff}$, the standard deviation, and whether they are pollinators or antagonists.

```{r}
stan_df = pardifs %>%
  group_by(species) %>%
  summarise(y=mean(dif), sigma=sd(dif)) %>%
  left_join(ecology) %>%
  mutate(pollinator = str_detect(interaction, 'pollinator'),
         antagonist = !str_detect(interaction, 'Dead')) %>%
  select(-interaction) %>%
  arrange(y)

stan_df 
```

Now let's build an hierarchical model in Stan that assumes that the estimate for each species is result of their interaction type plus a random fluctuation to be estimated. An example can be found [here](https://devinincerti.com/2015/10/31/bayesian-meta-analysis.html).

```{stan output.var=meta_model}
data {
  int<lower=0> N; // number of species
  real y[N]; // estimated difference of plant/space
  real<lower=0> sigma[N]; // sd of estimated difference
  int pollinator[N]; // whether species is a pollinator
  int antagonist[N]; // whether species is antagonist
}
parameters {
  real mu; 
  real pol;
  real ant;
  real<lower=0> tau;
  real eta[N];
}
transformed parameters {
  real theta[N];
  for (j in 1:N)
    theta[j] = mu + tau * eta[j] + ant * antagonist[j] + pol * pollinator[j];
}
model {
  mu ~ std_normal();
  tau ~ std_normal();
  eta ~ std_normal();
  pol ~ std_normal();
  ant ~ std_normal();
  y ~ normal(theta, sigma);
}
generated quantities {
  real yrep[N]; //posterior predictive simulations
  yrep = normal_rng(theta, sigma);
}
```



Let's now run the model.

```{r}
stan_input = as.list(stan_df)
stan_input$N = nrow(stan_df)
#stan_input$y = (stan_input$y- mean(stan_input$y))/sd(stan_input$y)
str(stan_input)
```

```{r}
fit = sampling(meta_model,
               seed = 1986317581,
               data = stan_input, 
               iter = 2000, 
               chains = 4,
               cores = 1,
               refresh = 2000,
               control = list(adapt_delta = 0.98, max_treedepth = 17))
```
Let's now save the fitted model.

```{r}
saveRDS(fit,file = 'stan_hier.rds')
```

## Model check
```{r}
#shinystan::launch_shinystan(fit)
```

To check if the model is doing a good job in fitting observations, let's plot $\alpha_{diff}$ that was used as input data agains posterior predictive draws:

```{r}
pps = rstan::extract(fit, pars='yrep',permuted=TRUE) %>%
  (function(x)x[[1]]) %>%
  t %>%
  as.data.frame %>%
  mutate(species = stan_df$species) %>%
  gather(key='iteration',value='yrep',-species) %>%
  mutate(species = factor(species,
                          ordered = T,
                          levels = c(
                            'Anchylorhynchus_1'  ,
                            'Anchylorhynchus_2'  ,
                            'Anchylorhynchus_3'  ,
                            'M_bondari_1'        ,
                            'M_bondari_2'        ,
                            'M_ypsilon'          ,
                            'R_rectinasus_1'     ,
                            'Andranthobius_1'    ,
                            'Andranthobius_2'    ,
                            'C_decolor_1'        ,
                            'C_decolor_2'        ,
                            'C_impar'            ,
                            'D_polyphaga'        ,
                            'P_cocoseae'),
                          labels = c(
                            "italic(Anchylorhynchus~trapezicollis)~OTU~1" ,
                            "italic(Anchylorhynchus~trapezicollis)~OTU~2",
                            "italic(Anchylorhynchus~trapezicollis)~OTU~3" ,
                            "italic(Microstrates~bondari)~OTU~1"         ,
                            "italic(Microstrates~bondari)~OTU~2"         ,
                            "italic(Microstrates~ypsilon)"               ,
                            "italic(Remertus~rectinasus)"            ,
                            "italic(Andranthobius~bondari)~OTU~1"       ,
                            "italic(Andranthobius~bondari)~OTU~2"       ,
                            "italic(Celetes~decolor)~OTU~1"         ,
                            "italic(Celetes~decolor)~OTU~2"         ,
                            "italic(Celetes~impar)"                 ,
                            "italic(Dialomia~polyphaga)"             ,
                            "italic(Phytotribus~cocoseae)"              
                          ))) %>%
  group_by(species) %>%
  slice(1:500) %>%
  ungroup %>%
  mutate(species = fct_reorder(species, yrep))


pps
```


```{r}
p = ggplot(dif_sums) +
  geom_abline(intercept = 0,slope = 0, color = 'gray90') +
  geom_jitter(aes(x=species,y=yrep),data=pps,alpha=0.1) +
  geom_linerange(aes(x = species, ymin = l95, ymax = u95, color = interaction),
                 show.legend = F) +
  scale_color_manual(values = clrs_light, name = 'Interaction') +
  ggnewscale::new_scale_color()+
  geom_linerange(aes(x = species, ymin = l50, ymax = u50, color = interaction),
                 size=1.3,
                 show.legend = F) +
  geom_point(aes(x = species, y = ave, color = interaction),
             fill = 'white',
             stroke = 1.5,
             shape = 21,
             size = 1.5) +
  scale_color_manual(values = clrs, name = 'Interaction', guide = 'none') +
  scale_fill_manual(values = clrs_light, name = 'Interaction', guide = 'none') +
  #geom_errorbar(aes(x = species, ymin = ave, ymax=ave, color = interaction)) +
  #geom_text(aes(x=species,y=pos,label=species,hjust=just), parse=T, size = 3) +
  #guides(color = guide_legend(override.aes = list(size = 1.5))) +
  ylab(expression(alpha[diff])) +
  coord_flip()+
  xlab("OTU") +
  scale_x_discrete(labels = parse(text = levels(pps$species))) +
  ylim(c(-0.35,0.35)) +
  theme_few() +
  theme(legend.position = c(0.99,0.01),
        legend.justification = c(1,0),
        text=element_text(size=9),
        #axis.text.y = element_text(face = 'italic',hjust=1),
        legend.text = element_text(size=6),
        legend.key.size = unit(8, 'pt'),
        legend.title = element_text(size=7)) 
  
  
print(p)

ggsave('figures/posterior_predictions_meta.pdf',
       p,
       device='pdf',
       width=168,
       height=168*3/4,
       units = 'mm',
       useDingbats=F)

```

It seems we are inferring with a little more variance than the original data, but the model is capturing the correct posterior estimates.


Let's now plot the posterior estimates.

```{r}
p2 = summary(fit, pars=c('ant','pol'))$summary %>%
  as.data.frame() %>%
  rownames_to_column('parameter') %>%
  #mutate(parameter=c('expression(gamma[a])','expression(gamma[p])')) %>%
  ggplot() +
  geom_abline(intercept = 0,slope = 0, color = 'gray90') +
  geom_linerange(aes(x = parameter, ymin = `2.5%`, ymax = `97.5%`),
                 color='gray70',
                 show.legend = F) +
  geom_linerange(aes(x = parameter, ymin = `25%`, ymax = `75%`), 
                 color='gray50',
                 size = 1.3,
                 show.legend = F) +
  geom_point(aes(x = parameter, y = mean),
             color='gray50',
             fill = 'white',
             stroke = 1.5,
             shape = 21,
             size = 1.5) +
  #geom_errorbar(aes(x = parameter, ymin = mean, ymax=mean),
  #              color='gray50') +
  scale_x_discrete(labels=parse(text=c('gamma[ant]','gamma[pol]'))) +
  ylab(expression(Estimated~effect~on~alpha[diff])) +
  xlab("Parameter") +
  coord_flip() +
  ylim(c(-0.35,0.35)) +
  theme_few()+
  theme(legend.position='bottom',
        text = element_text(size=9),
        axis.text.y = element_text(size=9))

print(p2)
  
```

Pollination does not seem to have a significant effect, and antagonism has a small effect but nonsignificant if we consider the whole posterior interval. Maybe with this sample size we do not have enough power to detect differences? To test, let's relabel our input data and consider that the top 3 species are antagonists, the next 3 are brood pollinators and remaining seven are commensals. We will fit the model and plot posterior estimates.

```{r}
stan_fake = stan_input

stan_fake$pollinator = c(rep(FALSE,7),rep(TRUE,3),rep(FALSE,3))
stan_fake$antagonist = c(rep(FALSE,7),rep(TRUE,6))

fake.fit = sampling(meta_model,
               seed = 9761,
               data = stan_fake, 
               iter = 2000, 
               chains = 4,
               cores = 1,
               refresh = 2000,
               control = list(adapt_delta = 0.98, max_treedepth = 17))
```

Let's now save the fitted model.

```{r}
saveRDS(fit,file = 'stan_hier_fake.rds')
```

And check the run with shinystan. This is commented out in the notebook since results can only be seen when running interactively.

```{r}
#shinystan::launch_shinystan(fake.fit)
```

It seems that with the relabelled data we have power to distinguish coefficients from 0:

```{r}
bayesplot::mcmc_intervals(as.array(fake.fit), pars = c('pol','ant'))
```


# Meta-analysis figures

Let's now prepare a final figure for the meta-analysis results, using cowplot. We will combine it with the color key manually later by using Adobe Illustrator.

```{r}
p = cowplot::plot_grid(p1,
                       p2,
                   ncol=1,
                   rel_heights = c(0.7,0.2),
                   align='v',
                   axis=c('lr'),
                   labels = 'auto',
                   label_y = c(1,1.5))

print(p)

ggsave('figures/meta_results.pdf',
       p,
       device='pdf',
       width=168/2*3,
       height=120,
       units = 'mm',
       useDingbats=F)

```

Let's also export a figure for supplementary material with the relabelled data to test the power of the test.
```{r}
fake_dif_sums = dif_sums %>% arrange(ave)
fake_dif_sums$interaction = factor(c(rep("Dead tissue breeder", 7),
                              rep("Brood pollinator", 3),
                              rep("Live tissue breeder", 3)),
                              levels = c("Brood pollinator",
                                         "Live tissue breeder",
                                         "Dead tissue breeder"))

p1 = ggplot(fake_dif_sums) +
  geom_abline(intercept = 0,slope = 0, color = 'gray90') +
  geom_linerange(aes(x = species, ymin = l95, ymax = u95, color = interaction),
                 show.legend = F) +
  scale_color_manual(values = clrs_light, name = 'Interaction') +
  ggnewscale::new_scale_color()+
  geom_linerange(aes(x = species, ymin = l50, ymax = u50, color = interaction),
                 size=1.3,
                 show.legend = F) +
  geom_point(aes(x = species, y = ave, color = interaction),
             fill = 'white',
             stroke = 1.5,
             shape = 21,
             size = 1.5) +
  scale_color_manual(values = clrs, name = 'Interaction', guide = 'none') +
  scale_fill_manual(values = clrs_light, name = 'Interaction', guide = 'none') +
  #geom_errorbar(aes(x = species, ymin = ave, ymax=ave, color = interaction)) +
  #geom_text(aes(x=species,y=pos,label=species,hjust=just), parse=T, size = 3) +
  #guides(color = guide_legend(override.aes = list(size = 1.5))) +
  ylab(expression(alpha[diff])) +
  coord_flip()+
  xlab("OTU") +
  ylim(c(-0.35,0.35)) +
  theme_few() +
  theme(axis.text.y = element_blank(),
        legend.position = c(0.99,0.01),
        legend.justification = c(1,0),
        text=element_text(size=9),
        legend.text = element_text(size=6),
        legend.key.size = unit(8, 'pt'),
        legend.title = element_text(size=7)) 

print(p1)

p2 = summary(fake.fit, pars=c('ant','pol'))$summary %>%
  as.data.frame() %>%
  rownames_to_column('parameter') %>%
  #mutate(parameter=c('expression(gamma[a])','expression(gamma[p])')) %>%
  ggplot() +
  geom_abline(intercept = 0,slope = 0, color = 'gray90') +
  geom_linerange(aes(x = parameter, ymin = `2.5%`, ymax = `97.5%`),
                 color='gray70',
                 show.legend = F) +
  geom_linerange(aes(x = parameter, ymin = `25%`, ymax = `75%`), 
                 color='gray50',
                 size = 1.3,
                 show.legend = F) +
  geom_point(aes(x = parameter, y = mean),
             color='gray50',
             fill = 'white',
             stroke = 1.5,
             shape = 21,
             size = 1.5) +
  #geom_errorbar(aes(x = parameter, ymin = mean, ymax=mean),
  #              color='gray50') +
  scale_x_discrete(labels=parse(text=c('gamma[ant]','gamma[pol]'))) +
  ylab(expression(Estimated~effect~on~alpha[diff])) +
  xlab("Parameter") +
  coord_flip() +
  ylim(c(-0.35,0.35)) +
  theme_few()+
  theme(legend.position='bottom',
        text = element_text(size=9),
        axis.text.y = element_text(size=9))

print(p2)



p = cowplot::plot_grid(p1,p2,
                   ncol=1,
                   rel_heights = c(0.7,0.2),
                   align='v',
                   axis=c('lr'),
                   labels = 'auto',
                   label_y = c(1,1.5))

print(p)

ggsave('figures/meta_supplement.pdf',
       p,
       device='pdf',
       width=168,
       height=100,
       units = 'mm',
       useDingbats=F)

```
